+++ src/MicroOcpp_c.cpp
@@ -1,5 +1,5 @@
 // matth-x/MicroOcpp
-// Copyright Matthias Akstaller 2019 - 2023
+// Copyright Matthias Akstaller 2019 - 2024
 // MIT License
 
 #include "MicroOcpp_c.h"
@@ -44,19 +44,6 @@ void ocpp_loop() {
 /*
  * Helper functions for transforming callback functions from C-style to C++style
  */
-
-MicroOcpp::PollResult<bool> adaptScl(enum OptionalBool v) {
-    if (v == OptionalTrue) {
-        return true;
-    } else if (v == OptionalFalse) {
-        return false;
-    } else if (v == OptionalNone) {
-        return MicroOcpp::PollResult<bool>::Await();
-    } else {
-        MO_DBG_ERR("illegal argument");
-        return false;
-    }
-}
 
 std::function<bool()> adaptFn(InputBool fn) {
     return fn;
@@ -138,12 +125,12 @@ MicroOcpp::OnReceiveErrorListener adaptFn(OnCallError fn) {
     };
 }
 
-std::function<MicroOcpp::PollResult<bool>()> adaptFn(PollBool fn) {
-    return [fn] () {return adaptScl(fn());};
+std::function<UnlockConnectorResult()> adaptFn(PollUnlockResult fn) {
+    return [fn] () {return fn();};
 }
 
-std::function<MicroOcpp::PollResult<bool>()> adaptFn(unsigned int connectorId, PollBool_m fn) {
-    return [fn, connectorId] () {return adaptScl(fn(connectorId));};
+std::function<UnlockConnectorResult()> adaptFn(unsigned int connectorId, PollUnlockResult_m fn) {
+    return [fn, connectorId] () {return fn(connectorId);};
 }
 
 void ocpp_beginTransaction(const char *idTag) {
@@ -291,10 +278,10 @@ void ocpp_addMeterValueInput_m(unsigned int connectorId, MeterValueInput *meterV
     addMeterValueInput(std::move(svs), connectorId);
 }
 
-void ocpp_setOnUnlockConnectorInOut(PollBool onUnlockConnectorInOut) {
+void ocpp_setOnUnlockConnectorInOut(PollUnlockResult onUnlockConnectorInOut) {
     setOnUnlockConnectorInOut(adaptFn(onUnlockConnectorInOut));
 }
-void ocpp_setOnUnlockConnectorInOut_m(unsigned int connectorId, PollBool_m onUnlockConnectorInOut) {
+void ocpp_setOnUnlockConnectorInOut_m(unsigned int connectorId, PollUnlockResult_m onUnlockConnectorInOut) {
     setOnUnlockConnectorInOut(adaptFn(connectorId, onUnlockConnectorInOut), connectorId);
 }