+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Replace `PollResult<bool>` with enum `UnlockConnectorResult`
+
 ### Added
 
+- File index ([#270](https://github.com/matth-x/MicroOcpp/pull/270))
+- Config `Cst_TxStartOnPowerPathClosed` to put back TxStartPoint
+- Build flags for customizing memory limits of SmartCharging
 - Operation GetInstalledCertificateIds, UC M03 ([#262](https://github.com/matth-x/MicroOcpp/pull/262))
 - Operation DeleteCertificate, UC M04 ([#262](https://github.com/matth-x/MicroOcpp/pull/262))
 - Operation InstallCertificate, UC M05 ([#262](https://github.com/matth-x/MicroOcpp/pull/262))
@@ -15,7 +22,12 @@
 
 ### Fixed
 
+- Fix defect idTag check in `endTransaction`
 - Make field localAuthorizationList in SendLocalList optional
+- Update charging profiles when flash disabled (relates to [#260](https://github.com/matth-x/MicroOcpp/pull/260))
+- Ignore UnlockConnector when handler not set
+- Reject ChargingProfile if unit not supported
+- Fix building with debug level warn and error
 
 ## [1.0.1] - 2024-02-27